Refactor backend setup and add swagger dependency
