+++ backend/Contracts/LoginRequest.cs
@@ -1,0 +1,3 @@
+namespace Backend.Contracts;
+
+public sealed record LoginRequest(string Usuario, string Contrasena);
