+++ backend/Contracts/LoginResponse.cs
@@ -1,0 +1,3 @@
+namespace Backend.Contracts;
+
+public sealed record LoginResponse(int Id, string Usuario, string Nombre);
